+++ frontend/app/(home)/components/HomeHeader/HomeHeaderBackground.tsx
@@ -1,0 +1,13 @@
+import styles from "./HomeHeader.module.css";
+
+export const HomeHeaderBackground = (): JSX.Element => {
+  return (
+    <div className="relative overflow-visible h-0 z-[-1]">
+      <div
+        className={`bg-gradient-to-b from-[#6300FF] to-[#D07DF9] ${
+          styles["bg-slanted-upwards"] ?? ""
+        } w-screen h-[25rem] z-[-1]`}
+      ></div>
+    </div>
+  );
+};
