+++ frontend/app/(home)/components/HomeHeader/QuivrLogo.tsx
